Add Show method to PaymentCongrats (#2245)
* show method added
* javadoc added!
* PaymentCongrats changed for PaymentCongratsModel
