+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/payment_congrats/model/PaymentCongratsCurrency.java
@@ -1,4 +1,4 @@
-package com.mercadopago.android.px.internal.features.payment_congrats;
+package com.mercadopago.android.px.internal.features.payment_congrats.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
